fix: fix cr issue that nested useCallback and useState for widget rendering
