+++ packages/gi-sdk/src/components/index.ts
@@ -1,3 +1,5 @@
 export { GraphContainer } from './GraphContainer';
+export { ImplWidget } from './ImplWidget';
 
 export type { GraphContainerProps } from './GraphContainer';
+export type { ImplWidgetProps } from './ImplWidget';
